technical improvement added typescript for this project
