--- src/components/ConfirmationModal/ConfirmationModal.js
@@ -1,43 +1,0 @@
-import React from 'react';
-import './ConfirmationModal.css';
-
-const ConfirmationModal = ({ 
-  isOpen, 
-  onConfirm, 
-  onCancel, 
-  title = 'Confirm Action', 
-  message = 'Are you sure you want to proceed?',
-  confirmText = 'Confirm',
-  cancelText = 'Cancel'
-}) => {
-  if (!isOpen) return null;
-
-  return (
-    <div className="modal-overlay" onClick={onCancel}>
-      <div className="modal-content" onClick={(e) => e.stopPropagation()}>
-        <div className="modal-header">
-          <h3>{title}</h3>
-        </div>
-        <div className="modal-body">
-          <p>{message}</p>
-        </div>
-        <div className="modal-footer">
-          <button 
-            className="btn btn-secondary" 
-            onClick={onCancel}
-          >
-            {cancelText}
-          </button>
-          <button 
-            className="btn btn-danger" 
-            onClick={onConfirm}
-          >
-            {confirmText}
-          </button>
-        </div>
-      </div>
-    </div>
-  );
-};
-
-export default ConfirmationModal;
