--- src/components/LoadingSpinner/LoadingSpinner.js
@@ -1,13 +1,0 @@
-import React from 'react';
-import './LoadingSpinner.css';
-
-const LoadingSpinner = ({ size = 'medium', message = 'Loading...' }) => {
-  return (
-    <div className="loading-container">
-      <div className={`spinner spinner-${size}`}></div>
-      <p className="loading-message">{message}</p>
-    </div>
-  );
-};
-
-export default LoadingSpinner;
