--- src/components/TodoForm/TodoForm.js
@@ -1,43 +1,0 @@
-import React, { useState } from 'react';
-import './TodoForm.css';
-
-const TodoForm = ({ onAddTask, isLoading = false }) => {
-  const [title, setTitle] = useState('');
-
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    if (title.trim()) {
-      onAddTask(title);
-      setTitle('');
-    }
-  };
-
-  return (
-    <form className="todo-form" onSubmit={handleSubmit}>
-      <div className="form-group">
-        <input
-          type="text"
-          value={title}
-          onChange={(e) => setTitle(e.target.value)}
-          placeholder="What needs to be done?"
-          className="todo-input"
-          disabled={isLoading}
-          maxLength={100}
-        />
-        <button 
-          type="submit" 
-          className="btn btn-primary add-btn"
-          disabled={!title.trim() || isLoading}
-        >
-          {isLoading ? (
-            <span className="loading-text">Adding...</span>
-          ) : (
-            <span>Add Task</span>
-          )}
-        </button>
-      </div>
-    </form>
-  );
-};
-
-export default TodoForm;
